Always skip any authentication
I've tested that this works even if the authenticate_user! before_filter hasn't been declared. It won't make any
practical difference to whether the request to /warmup starts the app but it should help us debug an odd issue
where requests are coming in to panopticon as /artefacts/warmup

I've also bumped the version to 0.1.0 as skipping the before filter feels like a significant change.